[Precision Depth Alignment] Change the pad_value parameter of pad3d from float to double (#75970) (#76127)
* Change the pad_falue parameter of pad3d from float to double

* fix Pad3dInferMeta

* fix1

* fix2

* fix3

* fix

* fix op_translator

* fix Pad3dOpTranscriber

+++ paddle/phi/infermeta/unary.h
@@ -559,7 +559,7 @@ PADDLE_API void PadInferMeta(const MetaTensor& input,
 PADDLE_API void Pad3dInferMeta(const MetaTensor& x,
                                const IntArray& paddings,
                                const std::string& mode,
-                               float value,
+                               double value,
                                const std::string& data_format,
                                MetaTensor* out,
                                MetaConfig config = MetaConfig());
